fix wrong number interface for Liquidated Vault (#805)

+++ Jellyfish.API.Loan/VaultLiquidationBatch.cs
@@ -2,7 +2,7 @@
 
 public class VaultLiquidationBatch
 {
-    public decimal Index { get; init; }
+    public double Index { get; init; }
     public string[] Collaterals { get; init; } = Array.Empty<string>();
     public string Loan { get; init; } = string.Empty;
 }
